refactor: split tui/model.go into model, keys, views, tasks
Pure file-level reorganization with no logic changes. The Model struct,
types, and lifecycle methods stay in model.go; key handling moves to
keys.go; rendering/view methods move to views.go; task filtering,
cursor navigation, and counting methods move to tasks.go.
